rename + xmldoc
